Adding a link to GettingStarted
I am adding it manually for now, as we need to somehow point users to
it when publishing to gh-pages. Later we will need to convert MD to
html and integrate it into documentation building script, I’ll create a
ticket for that.

The link is pointing to master of parser repo instead of a local copy
of MD file as the users, which cloned the repo, should also be able to
see it and might have no MD previewer installed.